Lab 3 finished
