+++ lab7/src/cpe200/BinaryCalculator.java
@@ -2,21 +2,7 @@ package cpe200;
 
 import java.math.BigDecimal;
 
-public class BinaryCalculator {
-
-    private BigDecimal firstOperand;
-    private BigDecimal secondOperand;
-
-    public void setFirstOperand(IOperand operand) {
-        Binarycheck(operand);
-        setFirstOperand(new BigDecimal(Integer.parseInt(operand.getOperand(),2)));
-    }
-
-
-    public void setSecondOperand(IOperand operand) {
-        Binarycheck(operand);
-        setSecondOperand(new BigDecimal(Integer.parseInt(operand.getOperand(),2)));
-    }
+public class BinaryCalculator extends BaseCalculator {
 
     public String add() throws RuntimeException {
         if(getSecondOperand().intValue()<0||getFirstOperand().intValue()<0)throw new RuntimeException("ERROR");
@@ -46,25 +32,22 @@ public class BinaryCalculator {
         return Integer.toBinaryString(getFirstOperand().stripTrailingZeros().pow(getSecondOperand().intValue()).stripTrailingZeros().intValue());
     }
 
-    public BigDecimal getFirstOperand() {
-        return firstOperand;
-    }
-
-    public void setFirstOperand(BigDecimal firstOperand) {
-        this.firstOperand = firstOperand;
-    }
-
-    public BigDecimal getSecondOperand() {
-        return secondOperand;
-    }
-
-    public void setSecondOperand(BigDecimal secondOperand) {
-        this.secondOperand = secondOperand;
-    }
-
     private void Binarycheck(IOperand check) throws ArithmeticException {
         if(check.getOperand().matches("[01]+"));
         else throw new ArithmeticException("Not Binary");
+    }
+
+
+    @Override
+    public void setFirstOperand(IOperand operand) {
+        Binarycheck(operand);
+        setFirstOperand(new BigDecimal(Integer.parseInt(operand.getOperand(),2)));
+    }
+
+    @Override
+    public void setSecondOperand(IOperand operand) {
+        Binarycheck(operand);
+        setSecondOperand(new BigDecimal(Integer.parseInt(operand.getOperand(),2)));
     }
 
 }
